component hierarchy and app structure re ordering
