+++ react-app/src/Components/Dashboard/MapComponent.jsx
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import { Map, GoogleApiWrapper, Marker, InfoWindow } from 'google-maps-react';
-import { Link } from 'react-router-dom';
 
 import '../../Stylesheets/map.css';
 import { getEstablishmentsInBounds } from '../../Queries/API';
